src/third_party/blink/web_tests/fast/dom/shadow/caret-range-from-point-in-nested-shadow-expected.txt

CONSOLE WARNING: line 25: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
PASS internals.treeScopeRootNode(range.startContainer) is document
PASS internals.treeScopeRootNode(range.endContainer) is document
PASS successfullyParsed is true

TEST COMPLETE
